test(export): address the three CodeRabbit nitpicks on the winding tests
All three verified against the code before acting; all three were right.

1. Move frame.rs's test module to a sibling frame_tests.rs, included via
   `#[cfg(test)] #[path = ...] mod frame_tests;`. Tests were 113 of the
   file's 223 lines, and geom.rs/geom_tests.rs in this same crate already
   follow that convention.

2. `trailing_partial_triangle_is_left_alone` asserted the `tri_end`
   boundary for `to_yup_into` only; `to_yup_in_place` computes it
   independently. Now both. This is the same duplicated-copy trap the
   file exists to close, so the omission mattered.

3. obj.rs: `mesh_visible` gates on `!indices.is_empty()`, so a visible
   mesh may hold one or two indices; `mesh.indices[0..3]` could panic.
   The suggested `len() >= 3` predicate fixes the panic but not the
   second half: such a mesh still writes its VERTICES and emits no face
   (`chunks_exact(3)`), so `vert_base` advances and the first face's
   indices need not start at zero. The test now walks the same sequence
   the exporter walks and accumulates the offset.

Re-ran all four winding mutations after the file split — a refactor can
drop tests from the build without touching them. Still 5 fail / 139 pass
mutated, 144/144 clean. Ratchet 4/4.
